🎨 디바이스 로그 출력 구조를 개선.
주요 로직이 아니니 별도 파일로 분리.
